Add files via upload
Here I write a program that reads the input video and outputs a modified version of the same video in which the number of bricks left in each column is displayed as a number above each column. There are 13 columns, so your program needs to overlay the video with 13 numbers and update them as the bricks disappear.